Add files via upload
These should be the files for the term project, it's a good skeleton of the homepage and contact section. I will be working on optimizing it in terms of display aesthetics but I am uploading it here for you, if you want to add anything but more importantly so you can try adding your functionality into it. I tried and i think i hit it but just in case you get double-check it on top of your code. -Manny